pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mully-mouth"
version = "3.0.0"
description = "AI-first voice caddy for GS Pro golf simulator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.7.0",
    "elevenlabs>=0.2.0",
    "opencv-python>=4.8.0",
    "mss>=9.0.0",
    "pygetwindow>=0.0.9",
    "pynput>=1.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "imagehash>=4.3.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
mully-mouth = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
